Multiplayer basics implemented with head tracking
Multiplayer implemented with head tracking mapped to a simple cube
avatar. Networking is very basic but works fine to instantiate a new
player in real-time right when the game starts up.

Also made slight modifications to the quality and project settings to
try and prevent some of the crashing behavior from the cooperative dev
last wed. Launches in windowed mode now at a lower resolution allowing
the player to resize. This seems to be much more stable and will work
well for testing at least. We can refocus on it more later once we are
ready to do more testing in the HMD.