Added mic support (always on at the moment)
Microphone is detected and will pickup any sounds and transmit them to
all other players connected to the game (including the person that is
talking). This will be fixed in a later commit so that the audio only
transmits to the other players and does not feedback to the source
player over the network.