Added multiple mic activation behaviors
Push the T button to control it. Currently it's configured to behave as
a Hold to Talk feature where T is the keyboard key to hold down.

Push to talk
Hold to talk
Always on

Some funky behavior when transmitting the audio over the network. Need
to track that down, so it wont work perfect right now and may crash in
fact.